fixing bug with deleting users

+++ admin/includes/view_all_users.php
@@ -73,10 +73,15 @@ if (isset($_GET['change_to_subscriber'])) {
 }
 
 if (isset($_GET['delete'])) {
-    $the_user_id = $_GET['delete'];
-    $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
-    $delete_query = mysqli_query($connection, $query);
-    header("Location: users.php");
+    if (isset($_SESSION['user_role'])) {
+        if ($_SESSION['user_role'] == 'admin') {
+
+            $the_user_id = mysqli_real_escape_string($connection, $_GET['delete']);
+            $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
+            $delete_query = mysqli_query($connection, $query);
+            header("Location: users.php");
+        }
+    }
 }
 
 ?>
